Need guidance figuring out how to update the ROOT_QUERY array using update function
In the addTodo mutation, I'm adding the newly created todo to the array of objects after it has been normalized in the cache. Unfortunately, the UI doesn't update after it has been changed. Perhaps this is because we are creating a new array of items from the existing one, but I'm not entirely sure.